+++ test/human_lib_format_time2_spec.rb
@@ -1,8 +1,8 @@
 require 'rspec'
 require 'st_tools'
 
-describe 'Проверка методов StTools::Human.format_time2' do
-    it '.format_time2 (:ru, :human, :full)' do
+describe 'Проверка методов StTools::Human.format_time2 (:ru)' do
+    it '.format_time2 (:human, :full)' do
       StTools.configure { |config| config.locale = :ru }
       test = ::StTools::Human.format_time2(Time.now, :human, :full)
       expect(test).to match(/[а-я]{3,10}/)
@@ -10,7 +10,7 @@ describe 'Проверка методов StTools::Human.format_time2' do
       expect(test).to match(/\d{2}\:\d{2}\:\d{2}/)
     end
 
-    it '.format_time2 (:ru, :human, :full, god:false)' do
+    it '.format_time2 (:human, :full, god:false)' do
       StTools.configure { |config| config.locale = :ru }
       test = ::StTools::Human.format_time2(Time.now, :human, :full, god: false)
       expect(test).to match(/[а-я]{3,10}/)
@@ -18,18 +18,41 @@ describe 'Проверка методов StTools::Human.format_time2' do
       expect(test).to match(/\d{2}\:\d{2}\:\d{2}/)
     end
 
-    it '.format_time2 (:ru, :full, :full)' do
+    it '.format_time2 (:full, :full)' do
       StTools.configure { |config| config.locale = :ru }
       test = ::StTools::Human.format_time2(Time.now, :full, :full)
       expect(test).to match(/\d{2}\/\d{2}\/\d{4}/)
       expect(test).to match(/\d{2}\:\d{2}\:\d{2}/)
     end
 
-    it '.format_time2 (:ru, :full, :short)' do
+    it '.format_time2 (:full, :short)' do
       StTools.configure { |config| config.locale = :ru }
       test = ::StTools::Human.format_time2(Time.now, :full, :short)
       expect(test).to match(/\d{2}\/\d{2}\/\d{4}/)
       expect(test).to_not match(/\d{2}\:\d{2}\:\d{2}/)
     end
+end
+
+describe 'Проверка методов StTools::Human.format_time2 (:en)' do
+  it '.format_time2 (:human, :full)' do
+    StTools.configure { |config| config.locale = :en }
+    test = ::StTools::Human.format_time2(Time.now, :human, :full)
+    expect(test).to match(/[a-zA-Z]{3,10}/)
+    expect(test).to match(/\d{2}\:\d{2}\:\d{2}/)
+  end
+
+  it '.format_time2 (:full, :full)' do
+    StTools.configure { |config| config.locale = :en }
+    test = ::StTools::Human.format_time2(Time.now, :full, :full)
+    expect(test).to match(/\d{2}\/\d{2}\/\d{4}/)
+    expect(test).to match(/\d{2}\:\d{2}\:\d{2}/)
+  end
+
+  it '.format_time2 (:full, :short)' do
+    StTools.configure { |config| config.locale = :en }
+    test = ::StTools::Human.format_time2(Time.now, :full, :short)
+    expect(test).to match(/\d{2}\/\d{2}\/\d{4}/)
+    expect(test).to_not match(/\d{2}\:\d{2}\:\d{2}/)
+  end
 end